Add service import, archive option
Add command line options taking the name of a service provider module,
and the path to the archive for the specified service. Implement dynamic
loading of the service provider module and creation of the service from
the same. This allows the user to specify the service, which need not be
provided with the application.