Improve user experience when downloading
Add a status bar reporting the number of active network requests, or
number of results, and a progress bar to show when we are waiting on
network requests. Also disable the search bar while we are downloading.
This gives the user visible feedback when we are actively talking to the
network.