Add a CMakeLists.txt
Add a basic CMake "build" system. This isn't strictly required for
Kochipylli itself at this point, since we are dealing with pure Python
code, but some IDE's (notably, KDevelop) need something build-like in
order to create a project, and as it generates rules to pre-compile the
Python into bytecode, it's convenient for syntax checking without trying
to actually run Kochipylli.