Start implementing result viewing
Hook up notification when a result is selected, a slot for the same, and
a method in Service to get the image path for a result. This will
eventually be used to show fetched results; for now it only shows the
result thumbnail (which we also want, so users can see the thumbnail
enlarged to help decide whether or not to fetch a result).