Start implementing machinery to fetch result
Modify ResultList.resultRequested signal to have the result name and
fetch URL, instead of the less helpful QListWidgetItem. Connect to new
Service.requestResult method (which just requests the fetch URL and
hands off to dispatchResultRequest, which must be implemented by the
specific service implementation). Add related methods for fetching the
actual image and saving it to the temporary area. Finally, add logic to
update the result list when the image is ready, and to show the full
result image rather than the thumbnail when available.