Add method to set info text
Add a method to Service to set information text (in the information
pane) from a key/value pair on a result. Besides being a slight
convenience, the new method also sets an explicit indicator when the
information is not available, rather than leaving the label blank (also
important to clear the text as set by a previous result).