Don't scroll when adding/removing results
As best as I can tell, at least QListView (possibly only in Batched
layout mode) has what seems to be a bug, where the scroll range is
temporarily reduced to a much smaller value (probably the size of one
batch of items) when reflowing items due to an addition or removal,
which causes the view to scroll somewhere undesired and AFAICT cannot be
worked around with scrollTo[Item]. Work around this by recording
position changes to the scroll bar in a way that lets us guess if a
change is due to user action or the aforementioned behavior, such that
we can restore the previous position when the layout process is
sufficiently complete that the old position becomes available again.

This isn't (and can't be) perfect, as it flickers a bit during the
layout process (e.g. when downloading many items), but is much, MUCH
better than without the work-around.