Set up save location selector
Port KFileTree (which for some reason isn't wrapped, although we end up
changing it anyway) from C++ to our local Python. Port meta-model class
from k3b to Python KMetaModel (dropping - for now at least - flat nodes,
drag-and-drop, and MIME support). Modify KFileTree to use the new
KMetaModel so that the root node is visible and selectable (which is
important for our usage).