Add 'new folder' button
Add and implement a button to create new folders. The user will likely
want to do this at some point, and it is much more convenient for us to
provide this feature than requiring the user to use an external shell or
file manager.

Note: Right now there is no code to select the newly created folder.
There are two problems preventing this from working; first, that the
directory model needs to be forcibly and synchronously updated, which
I couldn't get to work, and second, it seems there is no way to set the
current combo box item (except to root items, which is not helpful for
us) other than user interaction.