Improve method of joining paths
Create pathCat helper function as a Qt version of os.path.join. This
takes any number of arguments and calls QDir.cleanPath on the result (to
remove extra '/'s and '.' elements), and replaces the previous method
which had bunches of "%1/%2" literals floating around the code. (The new
method does something similar internally and is almost certainly
slightly less efficient, but the code is much cleaner.)