Improve existing file handling
Add check in Service.saveToDisk if the destination already exists. Use
QCryptographicHash to test if it is the same file, and show an explicit
error to that effect if it is not. If it is the same file, accept the
existing destination as a successful save and remove the source.