Implement result sets
Add machinery to keep track of which specific search caused us to get a
result. Add and hook up UI to select such result sets. This makes
searching much more sane in the presence of large prior result sets
and/or exploratory results, by allowing the user to review only the
results relevant to a specific search, while still keeping the total set
of results around for later perusal.