Add 'execute search' ('go') button
Add a specific button to execute the search. This is partly because it
feels like there should be one, but also because it is more sensible for
search types with no inputs (e.g. 'new images'), and also provides a
hook to indicate to the user if a search is possible (e.g. if there is
validation of search input fields, especially 'is empty').