Convert tool bars to QAction
Convert I/O tool bar buttons to QActions's, and a QToolBar (rather than
QHBoxLayout) to hold them, which gives us auto-raise and text style for
lower cost (besides that they really should be QAction's anyway). DO
likewise with the navigation tool bar, and also remove the hooks to not
add the actions.