Add option to include saved results
Add a search option to include in search results items which have
already been saved to the archive. Note that this depends on the service
implementation to actually request loading from the archive when it
detects that a result is already in the database. (Probably this should
be refactored to move more of this code to the base Service, but not
doing that now...)