Fix scaling of result thumbnails
Explicitly resize result thumbnails to fit into the desired size, if
larger. This is supposed to be handled by QAbstractItemView.iconSize,
but for some reason doesn't seem to be working. Also it's much more
strongly desirable for memory usage when we're generating thumbnails
from full-sized saved results (to not keep the full result image in
memory), which wasn't as important for result thumbnails off the web
that were at most a little bit larger than desired size.